refactor(audio): extract audio engine boundary
Move the audio engine abstraction into a dedicated capability-oriented service and update test doubles so AudioManager has a smaller, clearer core.\n\nGenerated with Codebuff 🤖\nCo-Authored-By: Codebuff <noreply@codebuff.com>
