+++ Oak/Oak/Services/AudioManager.swift
@@ -3,68 +3,6 @@ import Combine
 import Foundation
 import os
 
-// MARK: - AudioEngineProtocol
-
-internal protocol AudioEngineProtocol {
-    var isRunning: Bool { get }
-    var outputChannelCount: AVAudioChannelCount { get }
-    var outputSampleRate: Double { get }
-    func setMixerVolume(_ volume: Float)
-    func attachAndConnect(_ node: AVAudioNode)
-    func detach(_ node: AVAudioNode)
-    func prepare()
-    func start() throws
-    func stop()
-    func pause()
-}
-
-// MARK: - AudioEngineAdapter
-
-internal final class AudioEngineAdapter: AudioEngineProtocol {
-    private let engine = AVAudioEngine()
-
-    var isRunning: Bool {
-        engine.isRunning
-    }
-
-    var outputChannelCount: AVAudioChannelCount {
-        engine.outputNode.outputFormat(forBus: 0).channelCount
-    }
-
-    var outputSampleRate: Double {
-        engine.outputNode.outputFormat(forBus: 0).sampleRate
-    }
-
-    func setMixerVolume(_ volume: Float) {
-        engine.mainMixerNode.outputVolume = volume
-    }
-
-    func attachAndConnect(_ node: AVAudioNode) {
-        engine.attach(node)
-        engine.connect(node, to: engine.mainMixerNode, format: nil)
-    }
-
-    func detach(_ node: AVAudioNode) {
-        engine.detach(node)
-    }
-
-    func prepare() {
-        engine.prepare()
-    }
-
-    func start() throws {
-        try engine.start()
-    }
-
-    func stop() {
-        engine.stop()
-    }
-
-    func pause() {
-        engine.pause()
-    }
-}
-
 // MARK: - AudioManager
 
 @MainActor
@@ -79,12 +17,12 @@ internal class AudioManager: ObservableObject {
     @Published var isPlaying: Bool = false
 
     private var audioPlayer: AVAudioPlayer?
-    private var audioEngine: (any AudioEngineProtocol)?
+    private var audioEngine: (any AudioEngineControlling)?
     private var audioNodes: [AVAudioNode] = []
     private let logger = Logger(subsystem: "com.productsway.oak.app", category: "AudioManager")
-    private let audioEngineFactory: () -> any AudioEngineProtocol
+    private let audioEngineFactory: () -> any AudioEngineControlling
 
-    init(audioEngineFactory: @escaping () -> any AudioEngineProtocol = { AudioEngineAdapter() }) {
+    init(audioEngineFactory: @escaping () -> any AudioEngineControlling = { AudioEngineAdapter() }) {
         self.audioEngineFactory = audioEngineFactory
     }
